scraping_imdb/scraping_imdb/requirements.txt
altair==4.2.2
appdirs==1.4.4
asttokens==2.2.1
attrs==22.2.0
Automat==22.10.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
brotlipy==0.7.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==39.0.1
cssselect==1.2.0
debugpy==1.5.1
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
executing==1.2.0
filelock==3.11.0
gitdb==4.0.10
GitPython==3.1.31
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.4.1
incremental==22.10.0
ipykernel==6.15.0
ipython==8.12.0
itemadapter==0.8.0
itemloaders==1.0.6
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.1
pandas==1.5.3
parsel==1.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
platformdirs==3.2.0
prompt-toolkit==3.0.38
Protego==0.2.1
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pydeck==0.8.1b0
PyDispatcher==2.0.5
Pygments==2.15.0
PyHamcrest==2.0.4
pymongo==4.3.3
Pympler==1.0.1
pyOpenSSL==23.1.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.0
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
rich==13.3.4
Scrapy==2.8.0
service-identity==18.1.0
setuptools==65.6.3
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
stack-data==0.6.2
streamlit==1.21.0
tldextract==3.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
Twisted==22.2.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
w3lib==1.21.0
watchdog==3.0.0
wcwidth==0.2.6
wheel==0.38.4
zipp==3.15.0
zope.interface==5.4.0
